+++ yellowstone-grpc-geyser/src/plugin/entry.rs
@@ -5,14 +5,14 @@ use {
         metrics::{self, PrometheusService},
         parallel::ParallelEncoder,
         plugin::message::{
-            Message, MessageAccount, MessageBlockMeta, MessageEntry, MessageSlot,
-            MessageTransaction,
+            Message, MessageAccount, MessageBlockMeta, MessageDeshredTransaction, MessageEntry,
+            MessageSlot, MessageTransaction,
         },
     },
     agave_geyser_plugin_interface::geyser_plugin_interface::{
         GeyserPlugin, GeyserPluginError, ReplicaAccountInfoVersions, ReplicaBlockInfoVersions,
-        ReplicaEntryInfoVersions, ReplicaTransactionInfoVersions, Result as PluginResult,
-        SlotStatus,
+        ReplicaDeshredTransactionInfoVersions, ReplicaEntryInfoVersions,
+        ReplicaTransactionInfoVersions, Result as PluginResult, SlotStatus,
     },
     std::{
         concat, env,
@@ -307,6 +307,24 @@ impl GeyserPlugin for Plugin {
         })
     }
 
+    fn notify_deshred_transaction(
+        &self,
+        transaction: ReplicaDeshredTransactionInfoVersions,
+        slot: u64,
+    ) -> PluginResult<()> {
+        self.with_inner(|inner| {
+            let ReplicaDeshredTransactionInfoVersions::V0_0_1(transaction) = transaction;
+
+            let message = Message::DeshredTransaction(MessageDeshredTransaction::from_geyser(
+                transaction,
+                slot,
+            ));
+            inner.send_message(message);
+
+            Ok(())
+        })
+    }
+
     fn account_data_notifications_enabled(&self) -> bool {
         true
     }
@@ -320,6 +338,10 @@ impl GeyserPlugin for Plugin {
     }
 
     fn entry_notifications_enabled(&self) -> bool {
+        true
+    }
+
+    fn deshred_transaction_notifications_enabled(&self) -> bool {
         true
     }
 }